Further development on game feel components
Incorporated an animation curve for the prototype crown so bounces are more satisfying. Furthermore, I imported to newer run prototype animation of Rin running. Lastly, I spent a lot of time trying to fine tuning movement so its satisfying but there is still more work to be done.

I will continue to experiment on camera sizing so characters don't appear too small but there is also enough room to move about in a level.