Proportional navigation properly implemented with minor *bugs*
I finally completed some further tweaking on the proportional navigation for the crown prototype and now the crown is able to be thrown while locking on to nearby characters. However, the lock on is not too intense as to stop you from being able to 'select' who you wish to possess by throwing the crown at different initial trajectories. I do have some magic numbers in my code so I will come back later to fix on that. Interestingly, I found that the best way to make the crown deflections off walls not seem unpredicatable and frustating is to actually have the calculated deflection velocity to slightly rotate towards a character that was being tracked down. Thus, deflections deflected into a more desirable way, and that doesn't present frustation or the sense of RNG for the player.

Although, on the other hand, there are some minor bugs. However, it is better to state this as niche cases where the crown doesnt act in a desirable way that could potentially seem unfair or buggy to the user. Like in some very small cases where the crown appears to deflect away from the player, or ignores sections of the ground, or can appear to track the wrong player. It should be stated that all of these bugs will be fixed, but not now, since it is in my best interest to further develop the Ai to see how my proportional navigation works with moving targets. However, this does not mean I wont fix a few bugs here and there for now (like collision detection).

Either way, I am very happy about the results and want to see how animations, particle effects, and game feel will be once I implement complimenting components into my game.