Patch towards Proportional Navigation on crown and crown collisions
Fixed contradictory if statements that wasn't applying assisted deflections in order to have the crown have its deflection vector rotate towards another character if it meets the certain requirements.

Furthermore, I named the variables more intuitively and also allowed the crown to ignore insignificant collisions such as if the crown was chasing down a player but managed to nick the collision of stairs along the way.

Values can and still need to be fined tuned in the future. However, important note. I fixed crown collisions so the crown no longer falls out of the map or makes inaccurate deflections that are caused when the crown is submerged in another collider.

Final note, there are some magic numbers laying about in the code. So I will fix that on a later date.