Large commit towards animations
This commit contains large steps towards the development of possession as I have now implemented the jump animations as well as mapping them towards different velocities that are experienced during the duration of a jump. Furthermore, I have incorporated the a prototype forward light attack animation for Rin. I believe it may be best to incorporate prototype animations first and then finalize them all later. However, I will not leave running and jumping animations as prototypes. I also did some experimenting with having only the possessed characters show their shadows and having crowns be placed on the actually possessed characters. I would wish to upload this commit as a patch but there is still more to be done for v0.2.3-alpha.1 (animation update)