Large feature update
So this is the animation update. I have incorporated many prototype fighting animations that can be triggered with the mouse and begun to lay the framework for hitboxes, hurtboxes, health and damage. I will need to fix up some code I have disabled for controller input, but after that. I plan on working more on being able to allow players to damage one another.