Implemented knockback into the game, however
This mechanic is not entirely finished yet and there appears to be some bugs. I come back and fix this later. Either way, this knockback mechanic has also given me a new tool called 'backgroundVelocities' which should help me in the future when it comes to handling wind or dodges