Half-way done completing the overshoot waypoints function
Once done, I should be able to implement the function into all my my jumping waypoints. Then I should be able to add the waiting waypoint which shouldn't be too difficult. However, maintaining this code seems be a nightmare. Although, this is natural, considering this is Ai that we had to incorporate kinematic equations of our own, and that this file is reaching 2000 lines -.-

Fortunately, Ai is the only most programming intensive thing I have in possession. I hope...