Large commit towards finalizing the Ai pathfinding feature with the addition of a prototype of the confirmed Miami beach level
This is a large commit towards finalizing the Ai for Possession because at this point, the Ai is capable of navigating the majority of obstacles within my game. There are nuanced cases in which the pathfinding fails. However, I may address these errors in the future because they are not too problematic. As of now, the Ai can handle static objects well, but I will need to produce another feature update to help the Ai process dynamic objects.

In this update, I will provide a build to allow players to test and give feedback of the Ai's pathfinding.

This commit is worth 6 days of work, consisting largely of bug fixing and correcting misguided equations. Their is also an abudance of unused code in some of my files as a result of me experimenting different solutions towards the problem at hand. I plan to trim out that unused code & provide further commenting in the future before I hand in my code for grading (This is an assignment after all).

I am proud to state the my project is fairly stable, and now I will turn my attention towards networking and more combat.