Players now both spawn once a match starts and are able to move and perform attacks that are registered by the other player
However, critiques would be sluggish movement due to a lack of interpolation/extrapolation of the other player's movement. Furthermore, falling animations are not registered or even appear on the other player's screen. Either way, this is a very good foundation for Possession networking, and I am bound to improve this networking as I do more research into making the other player's movement look smooth.

Final note, it did take some time to wrinkle out the bugs as I learnt more and more about how PhotonView worked.

You can check the development build for an alpha test of the networking.

[P.S. Photon server settings has a missing app ID in order to protect Possession's official servers. However, you are free to add your own photon app ID as that will not interfere with official Possession servers.]