Large commit towards adding comments in code and bug fixing
At this stage, I have just handed in my programming assignment in which my code is being reviewed as well as my testing tables to assess and provide me with grading that depends on how flexible and robust my program is. I am going to upload my testing table up later so you can get an idea of some of the tests I did.